MULES::limitSum: Added direct support for partially constrained outlets
limitSum now only adjusts the patch fluxes that are not fixed supporting
partially constrained outlets at which the flow of some of the phases might be
restricted, e.g. a particle filter.

Additionally limitSum is now applied to the bounded fluxes in bounded predictor
part of the phase-fraction solution in multiphaseEuler phaseSystem::solve,
providing better boundedness without the need for maxMin to be applied to the
phase-fractions after solution to ensure boundedness.  This change is needed to
handle the fact that the phase transport fluxes are not necessarily
divergence-free and hence can cause unboundedness at the upper-limit even when
upwind discretisation is applied and limitSum is necessary to ensure boundedness
of all phase-fractions by limiting the fluxes such that the phase-fractions sum
to 1.