MULES::limiter: Added optional globalBounds control
By default MULES limits to both local (monotonic) and global (e.g. 0-1) bounds.
The local bounding is not strictly necessary, particularly if an NVD or TVD
scheme is used which will provide loose monotonicity, so it is perfectly
reasonable to use MULES only to ensure global bounds only which is less
computationally intensive and requires fewer iterations of the limiter to
converge.  This mode of operation can now be selected by setting the new
optional globalBounds control to true (or yes or on).