foamMonitor: Update
This update removes foamMonitor's dependence on the deprecated/removed
'refresh' command in gnuplot.

The interactive behaviour of the script and the gnuplot window have also
been improved. The script now launches gnuplot in the background and
then exits, leaving the command-line usable. The gnuplot window remains
interactive after being opened and now exits completely when closed with
the mouse or by pressing 'q' or 'Alt-F4'. It no longer pops-up again
later, requiring the process to be identified and manually killed.

This change also makes the script use the default interactive (gnuplot)
terminal, rather than the x11 terminal. On most systems this will be a
qt or wxt terminal. This will improve how the plot looks and enable
controls such as zoom and grids amd save-to-file.

Resolves bug report https://bugs.openfoam.org/view.php?id=4221