multiphaseEuler::phaseSystem::solve: New predictor-corrector algorithm for particulate phases
To ensure that the phase-fraction of particulate phases are limited by the
physical packing limit it is necessary to limit the particulate phase flux with
respect to the mean flux using the global rather than local bounds to allow for
phase accumulation.  To be able to apply local bounds to at least the high-order
discretisation terms it has been necessary to split the phase-fraction solution
into a predictor/corrector sequence in which a low-order flux predictor with
global bounds is corrected with the high-order flux contribution with local
bounds.  This new generalised predictor/corrector algorithm is applied to all
phases unless globalBounds is set true in the MULES controls in which case the
correction is not necessary and the predictor can be executed directly with the
high-order fluxes, applying the global bounds.