multiphaseEuler::phaseSystem: Added optional phase-fraction clip control
    //- Optionally clip the phase-fractions
    //  between their physical limits and to sum to 1.
    //  Defaults to false
    Switch clip;

The default setting for clip is false (no) because the latest MULES and limitSum
now ensure boundedness more rigorously (provided the pressure equation is solved
to a tight enough tolerance) and clipping the phase-fractions is unlikely to be
needed but ensuring that they sum to 1 may be needed for very long runs with
small time-steps if cumulative round-off error causes drift.

For example in fvSolution specify:

    "alpha.*"
    {
        nAlphaCorr      1;
        nAlphaSubCycles 2;
        clip            yes;
    }

which enables phase-fraction clipping.