tutorials: Removed unnecessary 'application' entries and calls to getApplication
The getApplication function has been extended so that in the absence of
an 'application' entry in the system/controlDict, it now looks for the
presence of a 'solver' or 'regionSolvers' entry and returns foamRun or
foamMultiRun as appropriate. This removes the need for an 'application'
entry in the majority of cases, but retains backwards compatibility with
cases that do have such an entry.

The tutorials have been updated to remove all 'application' entries and
calls to getApplication. This simplifies configuration and scripting by
removing duplicate information and layers of indirection.

The only cases remaining that might still use their 'application'
entries are those of legacy (i.e., non-modular) solvers that do not have
an ./Allrun script. The foamRunTutorials function uses the 'application'
entry in these cases to determine what executable to run. These cases
can either retain their 'application' entry, or they can be given an
./Allrun script. There are only two tutorials and three tests in
OpenFOAM for which this is the case, so this is not a common scenario.
For consistency with other cases in OpenFOAM, these five tutorials and
tests have been given ./Allrun scripts, and 'application' entries have
therefore been removed from the entirety of OpenFOAM.