functionObjects::scalarTransport: Updated execute and write sequencing
The scalarTransport no longer executes at the start of the time-loop, it now
executes at the end of each time-step using the converged flux field from the
time-step.  The only disadvantage of this consistency change is that the
database writeObjects call at the end of the time-step occurs before the
scalarTransport execution so the updated scalar field is not yet available, to
avoid this problem the scalar field is now set to NO_WRITE and the
scalarTransport::write function explicitly writes the scalar field according to
the write controls specified for the functionObject.

Additionally the Crank-Nicolson support in the scalarTransport functionObject
has been rewritten to operate even at the first time-step and with sub-cycling
support and part of the Crank-Nicolson formulation is included in the optional
implicit predictor according to what is consistent with the Courant number based
stability limit to provide faster convergence of explicit correctors.