fvc::surfaceSum: Returns a VolInternalField rather than a VolField
fvc::surfaceSum generates an extensive vol/cell field for which extrapolating to
the boundaries is very unlikely to be physical or useful and in most cases the
boundary values were discarded.  fvc::surfaceSum now returns a VolInternalField
only without boundary values and the calling code can generate appropriate
boundary values if required, this avoids wasted time and effort generating
unused boundary values, particularly in parallel which required communication to
calculate the processor patch values.

fvc::surfaceIntegrate now also returns a VolInternalField rather than a VolField
to avoid the same cost and communication overhead when the boundary values are
not needed.

fvc::surfaceIntegrateExtrapolate is a new function returning a VolField with
extrapolated boundary values which can be used when they are required.