functionObjects::phaseScalarTransport: Updated execute and write sequencing
The phaseScalarTransport no longer executes at the start of the time-loop, it now
executes at the end of each time-step using the converged flux field from the
time-step.  The only disadvantage of this consistency change is that the
database writeObjects call at the end of the time-step occurs before the
phaseScalarTransport execution so the updated scalar field is not yet available, to
avoid this problem the scalar field is now set to NO_WRITE and the
phaseScalarTransport::write function explicitly writes the scalar field according to
the write controls specified for the functionObject.