foamMergeCase: Utility for creating minimal "variants" of other cases
This utility copies a case from a given source directory and merges
differences into it. Differences are specified in dictionaries laid out
on disk in the same way as in the source case. This allows for variants
of a case to be created without replicating common parts of the
configuration.

For example, if the source case is configured to use a RAS turbulence
model, but the variant is to run laminar, then a
system/momentumTransport.orig file can be created for the variant which
contains only the following:

    // Set the momentum transport model to laminar simulationType
    laminar;

    // Remove the (now) redundant RAS settings for clarity
    ~RAS;

Files specifying the differences are, by default, located in the current
case directory. These must have a .orig extension to prevent them from
being overwritten by the merge operation.

Optionally, additional variant case directories can be supplied, in
which case multiple variations will be applied successively to the
source case files.

A foamUnMergeCase script has also been added, which when specified with
identical arguments will reverse the merge operation and recover the
original configuration with only the variant files present. The intended
workflow is for foamMergeCase to be used within an ./Allrun script, and
foamUnMergeCase within the corresponding ./Allclean script.

The incompressbleVoF/damBreak*, compressibleVoF/damBreak* and
incompressibleFluid/pitzDailySteadyMapped* tutorials have been modified
to use this script. This provides example usage, and also removes some
duplicated specification.