nCorr,nAlphaCorr -> nCorrectors: The number of correctors control is now consistent between algorithms
nCorrectors is now used to specify the number of correctors in PISO, PIMPLE, VoF
and multiphaseEuler phase-fraction solution, populationBalance,
scalarTransport and phaseScalarTransport.

Additionally nAlphaSubCycles -> nSubCycles in the VoF and multiphaseEuler
solvers as this setting is within the "alpha.*" sub-directory in fvSolvers so
including "Alpha" in the name of the controls is historical and unnecessary e.g.

    "alpha.*"
    {
        nCorrectors     1;
        nSubCycles      2;
    }