zoneGenerator: New dynamic zone generation system
A zoneGenerator is a runtime selectable class which generates a zoneSet
comprising one or more of pointZone and/or cellZone and/or faceZone:

Class
    Foam::zoneGenerator

Description
    Abstract base class for all zoneGenerators, providing runtime selection

Class
    Foam::zoneSet

Description
    Zone container returned by zoneGenerator::generate

    Optionally contains any combination of point, cell or face zones generated
    by the zoneGenerator::generate function which may simply be the zones
    returned by a lookup from the lists of zones stored on the mesh or generated
    by the particular \c zoneGenerator from a geometric search for example.

    Having \c zoneSet optionally hold zones of more than one type is
    particularly useful for the generation of complex coupled or related zone
    systems, for example rotating regions requiring both the \c cellZone of the
    rotating cells and the faces of the outer moving/sliding boundary.

A list of zoneGenerators is constructed from the constant/zonesGenerator file by
the optional zonesGenerator meshObject and updated or regenerated following mesh
changes:

Class
    Foam::zonesGenerator

Description
    MeshObject zoneGeneratorList instantiated by polyMesh

    to provide optional dynamic zones which may be updated following mesh motion
    and/or topology change.

Alternatively zones can be pre-generated from a list of zoneGenerators supplied
to the new createZones utility:

Application
    createZones

Description
    Executes the zoneGenerators specified and writes the zones generated

Usage
    \b createZones [OPTION]

    Options:
      - \par -region \<name\>
        Specify an alternative mesh region.

      - \par -mesh \<name\>
        Specify an alternative mesh.

      - \par -dict \<filename\>
        Specify alternative dictionary for the zoneGenerators,
        defaults to system/createZonesDict

      - \par -clear
        Clear all registered zones before generating new zones.

It is planned that this flexible system of dynamic zone generation and update
will replace the current awkward topoSet utility based on set generation and
conversion to zones and in principle avoid the need for sets, hugely simplifying
the handling of zones.

The new zoneGenerators based on the current topoSet sources will be written and
committed over time until all the useful topoSet sources have been replaced and
made obsolete, ultimately deprecating the topoSet utility altogether.