multiphaseEuler::phaseSystemSolve: Added new implicit predictor explicit corrector algorithm
The implicit predictor explicit corrector phase-fraction solution algorithm
originally developed for the VoF solvers has now been extended for
multiphaseEuler to provide faster solution than using the standard explicit
sub-cycling approach when running at high Courant numbers.

Running at high Courant numbers reduces temporal accuracy but is particularly
useful when evolving the flow in complex mixer problems before running a more
accurate time-resolved simulation or for cases with localised regions of high
velocity, the temporal accuracy of which is not important for the overall flow.

The new algorithm is selected in the same manner as for the VoF solvers by
setting the optional MULESCorr control in the phase-fraction sub-dictionary of
the solvers list in fvSolution:

    MULESCorr   yes;

which will enable MULES limited explicit corrections to a bounded implicit
prediction based on upwinded fluxes.  The explicit corrections can be relaxed
before adding to the prediction using the optional MULESCorrRelax control e.g.:

    MULESCorrRelax 0.5;

It is generally advisable to relax the corrections for better stability and
smoother solutions, particularly at high Courant numbers. MULESCorrRelax
defaults to 0.5 which is likely to be effective up to a max Courant number of
~5.

The number of explicit corrections applied to the implicit prediction is set by
nCorrectors e.g.:

    nCorrectors     3;

Increasing nCorrectors provides greater accuracy but at additional computational
cost and may need to be increased as the max Courant number increases unless
temporal accuracy is not important, e.g. if the simulation is to provide an
approximate initial evolved flow.