compressible::cavitationModels: Optionally specify liquid index on construction
Sometimes the index of the liquid phase is known by the calling code.
This change allows calling code to specify the index of the liquid phase
on construction and prevent the need for the user to specify the name of
the liquid phase in the coefficients dictionary of the cavitation model.