multiphaseEuler: Phase change fvModels
Phase change processes in multiphaseEuler are now implemented as
fvModels. This includes heat-transfer and mass-diffusion limited
interfacial evaporation and condensation, dissolving, cavitation, wall
and third-phase-surface boiling, homogeneous nucleation, and
reaction-driven transfers. The available phase-change fvModels are:

      heatTransferLimitedPhaseChange - Model for heat-transfer limited
                                       phase change processes of pure
                                       phases or major components of
                                       phases

     massDiffusionLimitedPhaseChange - Model for mass-diffusion-limited
                                       phase change processes such as
                                       evaporation or condensation of a
                                       minor component of a gas, or
                                       dissolving a minor component
                                       of a liquid

           multiphaseEulerCavitation - Cavitation model

                         wallBoiling - Nucleate boiling model on a
                                       number of wall patches

                 phaseSurfaceBoiling - Nucleate boiling model on the
                                       surface of a third (solid) phase

           reactionDrivenPhaseChange - Model which simulates surface
                                       reactions by consuming species
                                       that stay on the surface at the
                                       rate at which the reaction
                                       creates them

             homogeneousCondensation - Model for the homogeneous
                                       nucleation of liquid droplets out
                                       of a gaseous mixture

    homogeneousLiquidPhaseSeparation - Model for the homogeneous
                                       nucleation of a solid particles
                                       or immiscible liquid droplets out
                                       of a liquid mixture

This change greatly reduces the overhead associated with introducing a
new form of phase change. A new phase-change model, in principle needs
only a single stand-alone fvModel. Previously, an entire derivation of
phase system was required. The fvModels structure also permits arbitrary
combinations of models to be in operation simultaneously, it facilitates
linearisation into any of the transport equations, and it allows the use
of source conditions to set the values of fields that are added or
removed.

Additionally, sub-model construction now checks the type of interface
specified by the user against types that are required and/or prohibited
by the model. If the type does not match then an error is generated.

For example, liquid_dispersedIn_gas_inThe_gas represents a model for the
gas "side" of the liquid-gas interface, when the liquid takes the form
of droplets dispersed within the gas. It is not correct to specify such
an interface for (say) a drag model, as the drag model has no dependence
on the "side" of the interface. An error to this effect is now generated
if this interface is used for a drag model. Previously the code just
silently removed the "inThe_gas" part of the specification.

In order to faciliate these changes, some changes to the syntax of the
wider multiphaseEuler setup have been made:

There is no selection of the phase system any more. The 'type' entry in
constant/phaseProperties has been removed. Its presence will result in a
warning.

Momentum transfer modelling (i.e., drag, virtual mass, lift, wall
lubrication and turbulent dispersion) has been removed from
constant/phaseProperties and is now specified in a new
constant/momentumTransfer file. E.g.:

    drag
    {
        blending        segregated;

        air_dispersedIn_water
        {
            type            SchillerNaumann;
        }

        water_dispersedIn_air
        {
            type            SchillerNaumann;
        }

        air_segregatedWith_water
        {
            type            segregated;
            m               0.5;
            n               8;
        }
    }

    virtualMass
    {
        air_dispersedIn_water
        {
            type            constantCoefficient;
            Cvm             0.5;
        }

        water_dispersedIn_air
        {
            type            constantCoefficient;
            Cvm             0.5;
        }
    }

    lift
    {}

    wallLubrication
    {}

    turbulentDispersion
    {}

Similarly, heat transfer modelling has been moved to
constant/heatTransfer. For example:

    blending
    {
        type            linear;
        minPartlyContinuousAlpha.air 0;
        minFullyContinuousAlpha.air 1;
        minPartlyContinuousAlpha.water 0;
        minFullyContinuousAlpha.water 1;
    }

    air_dispersedIn_water
    {
        type            RanzMarshall;
        residualAlpha   1e-4;
    }

    water_dispersedIn_air
    {
        type            RanzMarshall;
        residualAlpha   1e-4;
    }

Note that in this case a heatTransfer sub-dictionary is not required as
this file only contains heat transfer modelling. The new format also
allows for both one- and two-resistance modelling to be specified
simultaneously. If this is done then care must be taken to ensure that a
given interface does not have a mixture of one- and two-resistance
modelling specified.

The blending approach can now also be specified within the model
specification (see constant/heatTransfer above), or it can refer to the
blending section of constant/phaseProperties by name (see
constant/momentumTransfer/drag above). If blending is not included in
he model specification, then the model name or 'default' is looked up
in constant/phaseProperties/blending. This fallback is equivalent to the
existing behaviour.

These changes are backwards compatible. Momentum and heat transfer
modelling will continue to be read from constant/phaseProperties if the
constant/momentumTransfer and/or constant/heatTransfer files do not
exist. A warning will be generated to indicate if this is the case.

Entries in constant/phaseProperties files relating to phase change
processes - i.e., phaseTransfer, saturationTemperature,
interfaceComposition and diffusiveMassTransfer - are no longer supported
and their presence will result in an error.