refineMesh: Added direct support for hexRef8 refinement to replace refineHexMesh
By adding support for hexRef8 into the refineMesh utility it is now possible to
refine a zone or list of zones in a single call and output the cellLevel and
pointLevel files to support further refinement by refineMesh, snappyHexMesh or
dynamic refinement/unrefinement.

This new option is selected by the optional hexRef8 switch which defaults to
false to provide backward-compatibility, e.g. to 2x2x2 refine the hex cells in a
single cylindrical region the following entries in the refineMeshDict file are
required:

hexRef8         yes;

zone
{
    type    cylinder;
    point1  (0 0 0.5);
    point2  (0 0 0.7);
    radius  0.25;
}

or to refine the cells in a box and again in a slightly smaller box:

hexRef8         yes;

zones
{
    level1
    {
        type    box;
        box     (-1e6 -1e6 -1e6)(1e6 -0.2 0.35);
    }

    level2
    {
        type    box;
        box     (-1e6 -1e6 -1e6)(1e6 -0.15 0.3);
    }
}

For more details of the options in refineMesh see
etc/caseDicts/annotated/refineMeshDict

Also the deprecated autoRefineMesh utility has been removed as it has been
superseded by snappyHexMesh.