subsetMesh: Added support for sub-setting the mesh including the cells in a cellZone
The original cellSet argument has been replaced with a -cellSet option followed
by the cellSet name or the new -cellZone option followed by the cellZone name
can be used.

If the cellSet option is used the set of cells to keep in the sub-setted mesh is
read from the corresponding cellSet file.

If the cellZone option is used the set of cells to keep in the sub-setted mesh is
obtained from the corresponding cellZone stored on the mesh.

Alternatively the new optional subsetMeshDict file can be provided in the system
directory or another dictionary specified using the new -dict option which
contains the zone and optional patch specification, e.g. for the damBreak3D case:

    zone
    {
        type invert;

        box
        {
            type        box;
            zoneType    cell;
            box         (0.375 0.375 0) (0.625 0.625 0.25);
        }
    }

    patch   walls;

If the zone has been previously created, e.g. using createZones it can either be
specified using the new -cellZone option or in the subsetMeshDict file:

    zone <zone name>;

    patch   walls;

In the hotBoxes case the set of cells is obtained by inverting the cells in a
list of 4 boxes:

    zone
    {
        type invert;

        box1
        {
            type        box;
            zoneType    cell;
            box         (0.4 0.1 0.1) (0.6 0.3 0.3);
        }

        box2
        {
            type        box;
            zoneType    cell;
            box         (0.4 0.1 0.4) (0.6 0.3 0.6);
        }

        box3
        {
            type        box;
            zoneType    cell;
            box         (0.4 0.4 0.1) (0.6 0.6 0.3);
        }

        box4
        {
            type        box;
            zoneType    cell;
            box         (0.4 0.4 0.4) (0.6 0.6 0.6);
        }
    }

    patch   film;