populationBalance: Improve syntax and simplify code structure
Parameters relating to the groups' geommetries are now specified for the
population balance as a whole, rather than in a separate sub-dictionary
for each group. This information is usually just the groups'
representative spherical diameters, but for non-spherical shape models
(i.e., fractal) other properties relating to the groups' geommetries are
also required. For the diameters, typically, a minimum and a maximum are
specified along with an option for how intermediate values are to be
distributed. Other properties are given as Function1-s of the diameter.
For example, in constant/phaseProperties:

    air1
    {
        type            pureIsothermalPhaseModel;
        diameterModel
        {
            type            populationBalance;

            // Define the name of the population balance model
            // associated with this phase
            populationBalanceModel bubbles;

            // Set the number of groups assigned to this phase
            nGroups         10;
        }
        residualAlpha   1e-6;
    }

    air2
    {
        type            pureIsothermalPhaseModel;
        diameterModel
        {
            type            populationBalance;
            populationBalanceModel bubbles;
            nGroups         5;
        }
        residualAlpha   1e-6;
    }

    ...

    // Settings for the 'bubbles' population balance model referred to
    // in the settings for phases air1 and air2
    bubbles
    {
        ...

        // Define the representative spherical diameters of the groups
        sphericalDiameters
        {
            // Options are 'equal', 'exponential', or 'manual' spacings
            type            exponential;

            // For 'equal' and 'exponential' the maximum and minimum
            // diameters must be provided
            min             2.0e-3;
            max             1.2e-1;

            // For 'manual' a list of values is given
            //values          (2.0e-3 2.1e-3 2.2e-3 ...);
        }

        // Select the shape model
        shapeModel      spherical;

        ...
    }

The 'velocityGroup' diameter model has been renamed 'populationBalance'
as it relates specifically to a population balance model and the
population balance approach in general. The name 'velocityGroup' will
remain selectable for backwards compatibility.

The 'shapeModel' is now specified in the population balance model's
sub-dictionary rather than in the sub-dictionaries of the phases.

The 'populationBalances' entry has been removed. It is not needed as the
system can figure out the complete list of population balance models
directly from the phases' diameter models.

The 'populationBalanceCoeffs' sub-dictionary has been made optional, and
its use is discouraged. Population balance models are somewhat
conceptually similar to phases, and their names must be similarly unique
so that properties like 'd.<populationBalanceName>' and 'd.<phaseName>'
can be distinguished. This means it is both natural and possible to
specify settings for population balance models at the top level of the
phaseProperties file alongside the phases.

A number of boundary and source conditions have been changed to be
consistent with the changes described above. Where a condition had
'SizeGroup' as part of it's name it now has 'GroupFraction', and where
one had 'SurfaceAreaVolumeRatio' it now has
'GroupSurfaceAreaVolumeRatio'. A list of affected conditions is given
below. In all cases the old name remains selectable for backwards
compatibility.

    singleGroupFraction
    distributionGroupFraction
    interfacialGrowthGroupFraction
    nucleationGroupFraction
    uniformFixedValueGroupSurfaceAreaVolumeRatio
    interfacialGrowthGroupSurfaceAreaVolumeRatio
    hardSphereGroupSurfaceAreaVolumeRatio
    ParkRogakGroupSurfaceAreaVolumeRatio
    nucleationGroupSurfaceAreaVolumeRatio

The population balance code structure has been simplified with the
removal of the 'sizeGroup' class. This class followed the
"model-is-a-field" pattern that has become increasingly discouraged over
time. In particular, the constraint that the size-group model (and
associated sub-models) relate to a single group, rather than a list of
groups, proved restrictive in hindsight. This constraint is also largely
what lead to the "sub-dictionary of parameters for every group" syntax,
so the change is closely coupled to the syntax improvements listed
above.