solvers::compressible.*VoF: Added support for pRef
The optional reference pressure pRef can be specified in the optional
constant/pRef file e.g.

    dimensions      [1 -1 -2 0 0 0 0];
    value           1e5;

and is used when calculating the static pressure from p_rgh:

    p = p_rgh + rho*buoyancy.gh + buoyancy.pRef;

The advantage of providing a pRef is that it allows for better solver control
and normalisation of p_rgh and reduces the round-off error when the dynamic
range of p_rgh is small compared to the average level.

The same approach is used in the multiphaseEuler solver.