faceZone: The orientation flipMap is now optional
If the flipMap is not set the faceZone is considered to be unoriented and the
new member function oriented() returns false and the flipMap is not written into
the faceZones file for the corresponding faceZone.

The combinatorial operations union, difference, invert and intersection take
into account the orientation of the faceZones and combine the flipMaps where
possible or warn if the combination would produce an incomplete flipMap.