src/parallel/parallel: Use field decomposers and reconstructors without write
This change moves the field decomposer and reconstructor classes from
decomoposePar and reconstructPar into the parallelisation library. It
also adds public methods which return the decomposed/reconstructed field
to the calling code rather than writing them to disk. This enables
these classes to be used for on-the-fly decomposition and reconstruction
purposes.