paraFoam: Enable reading of decomposed cases
A new 'Decomposed Case' checkbox control has been added to the
properties panel in paraFoam. When checked, this makes paraFoam read the
decomposed case; i.e., the mesh and fields in the processor directories,
rather than those in the top-level case directory. This allows the case
to be visualised without having to run reconstructPar first.

paraFoam achieves this by reconstructing the case on-the-fly, using the
same reconstruction routines as are used by reconstructPar. Both the
fields and the mesh will be reconstructed as necessary.

These reconstruction processes have a cost associated with them. As a
result it is still likely to be beneficial to run reconstructPar in
advance of paraFoam in a number of cases; e.g., situations in which the
mesh reconstruction is of significant expense, or when the intention is
to visualise a large proportion of the available fields or time
directories. Conversely, on-the-fly reconstruction is likely to be
preferable when the reconstructed mesh is already available (or
relatively quick to reconstruct) and only one or two fields are of
interest at an intermediate time.

It is therefore anticipated that this new functionality will primarily
be of benefit when monitoring a running case, rather than doing the
final post-processing of a completed case.

ParaView's built in reader module (paraFoam -builtin) also has the
ability to read a decomposed case, but it operates somewhat differently.
It does not formally reconstruct the case. Instead it loads the
processor cases separately and combines them into a single dataset
without connecting them along their processor boundaries (this can be
observed by using the 'Connectivity' filter). This approach is likely to
be quicker (particularly if the reconstructed mesh is not available),
but the connectivity of the result is not strictly correct, and this may
affect the results of subsequent visualisation operations.