functionObjects::patchCutLayerAverage: Added 'distance' option
This option lets the user specify the name of a scalar distance field,
contours of which are used to create the layers across which the
averaging is performed. Either this option or the existing 'direction'
option should be provided to define the layers; not both. The distance
field may be a pointScalarField or a volScalarField. If it is the latter
it will be interpolated to the points.