ZoneList: Added optionally generated and cached "all" zone
which holds an identity map of all the points, cells or faces.  This can be
looked-up using the lookup zoneGenerator and used in any context supporting
zoneGenerators, in particular fvCellZone, which now using the ZoneList::all()
instead of the thread unsafe static identity map it used to use.  This changes
has allowed a simplification and rationalisation of all uses of fvCellZone.