functionObjects: Added cutLayerAverage
This function object writes graphs of cell values, volume-averaged in
planes perpendicular to a given direction or in contours of a given
distance field. It adaptively grades the distribution of graph points to
match the resolution of the mesh.

Unlike the layerAverage function, this function does not require the
mesh to be of any specific structure. This function will cut cells that
span multiple layers in order to distribute their contributions between
the layers.

Example of function object specification:

    cutLayerAverage1
    {
        type            cutLayerAverage;
        libs            ("libcutLayerAverageFunctionObject.so");

        writeControl    writeTime;
        writeInterval   1;

        direction       (1 0 0);
        nPoints         100;
        interpolate     no;

        fields          (p U);

        axis            x;
        setFormat       raw;
    }

A packaged function object is also included, which permits the following
syntax to be used, either with #includeFunc in system/functions, or with
the -func option to foamPostProcess:

    cutLayerAverage
    (
        funcName=graphX,
        direction=(1 0 0),
        nPoints=100,
        p,
        U
    );