aeratedStirredTankMRF: representative case of a stirred tank reactor, running with the multiphaseEuler solver.
This commit adds a representative industrial case for simulating an
aerated stirred tank, operating with air as the gaseous phase and water
as the liquid phase. The tank is baffled and the stirrer features a
pitch-blade impeller at the top and a Rushton turbine at the bottom. The
geometric ratios adhere to the recommendations given by Middleton &
Smith [^1], making the flow regime map published therein applicable.

The setup was tested for various combinations rotational speeds and gas
flow rates, resulting in the typical flow regimes occurring in aerated
stirred tanks:

| regime                                                    | rpm | m^3/
|                                                           |     | min
| --------------------------------------------------------- | --- | ----
| fully flooded stirrer, no dispersion                      | 100 | 0.37
| gas dispersed in upper part, small gas cavities at blades | 200 | 0.15
| gas dispersed in upper part, large gas cavities           | 200 | 0.75
| complete dispersion, small gas cavities                   | 500 | 0.37
| complete dispersion, large gas cavities                   | 500 | 1.12

The case uses the semi-implicit MULES algorithm for solving the phase
fraction transport equation. The maximum Courant number is set to a high
value of 40 to quickly reach a steady-state for the gas holdup, which
should be reduced for further more accurate transient simulations.

[^1]: Middleton, J. C., & Smith, J. M. (2004). Gas–Liquid Mixing in
Turbulent Systems. In: Handbook of industrial mixing.

Initial case created by CFD Direct, developed further by Institute of
Fluid Dynamics, Helmholtz-Zentrum Dresden - Rossendorf (HZDR)