etc/config: Control ThirdParty MPI verson via prefs.sh
The ThirdParty OpenMPI and MPICH versions are no longer hard coded
inside the configuration. They are now controlled by OPENMPI_VERSION and
MPICH_VERSION environment variables, which can be set within a
~/.OpenFOAM/<version>/prefs.sh file or similar. If these variables are
not set then the configuration will search for and use the ThirdParty
source directory with the newest version number.