paraFoam: Enforce 'snap-to-timestep' mode at all times
This prevents ParaView from entering 'sequence' mode if only one time
directory exists on disk. In this mode a number of times are seemingly
made available when they are in fact not present. This is counter
intuitive. It is also moderately difficult to get out of this mode if
times are subsequently generated.