zoneSet,zoneGenerators::lookup: Improved zone access checking and diagnostic messages
If the requested zone with the specified zoneType is not available lookup now
prints if it available with a different type, e.g. if in the ballValve
createZonesDict the lookup of the ball is specified to be of faceZone type:

ball
{
    type        point;
    ball
    {
        zoneType face;
    }
}

the error message

--> FOAM FATAL IO ERROR:
Cannot find faceZone ball
    Found cellZone ball
    Available faceZones: 0()

file: .../tutorials/incompressibleFluid/ballValve/system/createZonesDict/ball at line 21.

is printed.