surfaceFieldValue, volFieldValue: De-clutter the log
The number of faces, area, volume, etc..., are no longer automatically
printed into the log on mesh change. Instead the complete configuration
and the extent of the calculation region is printed only on read. After
that, the extent is treated like a field. It can be written into the
output file using the new write flags 'writeArea', 'writeNumberOfFaces',
'writeVolume' or 'writeNumberOfCells' as appropriate. It can also be
written into the log by setting both the write flags and the 'log'
control to true.

This change vastly de-clutters the log in the case when many field value
functions are present and the mesh is continuously changing.