functionObjects::sampledSets: Added support for dictionary format for sets specification
It is now possible to specify the sets list as a dictionary:

sets
{
    data
    {
        type    lineFace;
        axis    x;
        start   (-4.995 0 0);
        end     (4.995 0 0);
        nPoints 1000;
    }
}

or the previous list format:

sets
(
    data
    {
        type    lineFace;
        axis    x;
        start   (-4.995 0 0);
        end     (4.995 0 0);
        nPoints 1000;
    }
);

The dictionary format is more consistent with most other input in OpenFOAM and
allows lookup and substitution from other parts of the dictionary or other
files.

All tutorials have been updated to use the dictionary format.