paraFoamServer: Script to launch ParaView Server (pvserver) for OpenFOAM
This script launches ParaView Server (pvserver) with the OpenFOAM reader
module libraries. It forms the basis of a workflow in which an OpenFOAM
installation exists only on a centralised computing resource. It allows
OpenFOAM results to be post-processed on workstations without an
OpenFOAM installation and without transferring and copying the data from
the server.

Connecting to a ParaView Server instance also requires a client side
server configuration file (a .pvsc file). Details of how to construct
such a file can be found within ParaView's documentation.